% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_code_hierarchy.R
\name{analyze_code_hierarchy}
\alias{analyze_code_hierarchy}
\title{Analyze Parent-Child Code Hierarchies}
\usage{
analyze_code_hierarchy(
  data_inputs,
  dictionary,
  sample_labels,
  phecodes,
  custom_children = NULL
)
}
\arguments{
\item{data_inputs}{Named list of cleaned codified datasets.}

\item{dictionary}{Cleaned feature dictionary.}

\item{sample_labels}{Named list labeling each sample.}

\item{phecodes}{Character vector of parent Phecodes to analyze.}

\item{custom_children}{Optional named list of child codes for each parent.}
}
\value{
A list of plots for each Phecode, including `rate_plot` and `combined_plot`.
}
\description{
Evaluates hierarchical consistency for selected Phecodes by comparing trends for parent and child codes.
}
