% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_cui_alignment.R
\name{code_cui_alignment}
\alias{code_cui_alignment}
\title{Assess Code-CUI Temporal Alignment}
\usage{
code_cui_alignment(
  data_inputs,
  target_code,
  target_cui,
  dictionary,
  sample_labels
)
}
\arguments{
\item{data_inputs}{Named list of codified and NLP datasets.}

\item{target_code}{The Phecode of interest.}

\item{target_cui}{The CUI of interest.}

\item{dictionary}{Cleaned feature dictionary.}

\item{sample_labels}{Named list labeling the samples.}
}
\value{
A list with plots: `rates_plot`, `counts_plot`, and `correlation_plot`.
}
\description{
Analyzes agreement between a codified code and its corresponding CUI over time. Produces trends, patient counts, and correlation plots.
}
