% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{extract_data_for_visualization}
\alias{extract_data_for_visualization}
\title{Extract Data for Visualization}
\usage{
extract_data_for_visualization(
  sqlite_file,
  prefix,
  top_n = 20,
  additional_vars = NULL,
  wanted_items_df = NULL,
  manual_replacement_bank = NULL,
  dict_prefix = prefix
)
}
\arguments{
\item{sqlite_file}{Path to the SQLite file containing processed data.}

\item{prefix}{Prefix for medical codes (e.g., "RXNORM:", "LOINC:").}

\item{top_n}{Number of top occurrences to select (default = 20).}

\item{additional_vars}{A vector of additional variables to include (default = NULL).}

\item{wanted_items_df}{A data frame of specific variables to extract (default = NULL).}

\item{manual_replacement_bank}{A named list for renaming specific codes (default = NULL).}

\item{dict_prefix}{Dictionary prefix for mapping (default = `prefix`).}
}
\value{
A list containing:
  - `Total_Counts`: A summary of total occurrences for each medical code.
  - `Patient_Counts`: The number of unique patients associated with each medical code.
}
\description{
This function extracts and processes data from an SQLite database for visualization.
}
