% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{extract_patient_counts_over_years}
\alias{extract_patient_counts_over_years}
\title{Extract Patient Counts Over Time}
\usage{
extract_patient_counts_over_years(
  sqlite_file,
  codes_of_interest,
  dictionary_mapping
)
}
\arguments{
\item{sqlite_file}{A string path to the intermediary SQLite database file generated from Step 1.}

\item{codes_of_interest}{A character vector of full codes (e.g., "RXNORM:1234", "PheCode:714.1") to extract.}

\item{dictionary_mapping}{A data frame that maps codes to descriptions. It must include `Group_Code`, `Group_Description`,
`Common_Ontology_Code`, and `Common_Ontology_Description`.}
}
\value{
A named list of data frames:
* Each element corresponds to one code, showing `Year`, `Parent_Code`, `Patient_Count`, and `Name`.
* The `"combined"` element contains the merged data for all codes.
}
\description{
Extracts patient counts for specified codes from a time-aggregated intermediary SQLite file. 
For each code selected, it returns the number of unique patients observed in each year.
}
