% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{generate_intermediary_sqlite}
\alias{generate_intermediary_sqlite}
\title{Generate Intermediary SQLite Database}
\usage{
generate_intermediary_sqlite(con, output_sqlite_path)
}
\arguments{
\item{con}{A database connection object to an existing SQLite database.}

\item{output_sqlite_path}{Path to the output SQLite file where processed data will be stored.}
}
\value{
Saves the aggregated data to an SQLite file.
}
\description{
This function aggregates patient data and stores it in an intermediary SQLite file.
}
