% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_related_features.R
\name{plot_related_features}
\alias{plot_related_features}
\title{Visualize ONCE-Derived Related Features}
\usage{
plot_related_features(
  data_inputs,
  target_code,
  target_cui,
  sample_labels,
  O2 = FALSE,
  manual_ONCE_path_code = NULL,
  manual_ONCE_path_nlp = NULL,
  types = c("Diagnosis", "Medication", "Lab", "Procedure", "CUI"),
  type_dict = list(Diagnosis = "PheCode", Medication = "RXNORM", Lab = c("LOINC",
    "ShortName", "Other Lab"), Procedure = "CCS")
)
}
\arguments{
\item{data_inputs}{Cleaned codified and NLP datasets.}

\item{target_code}{Target Phecode.}

\item{target_cui}{Target CUI.}

\item{sample_labels}{Named list labeling the samples.}

\item{O2}{Logical flag for using O2 default paths.}

\item{manual_ONCE_path_code}{Optional codified dictionary path if O2 = FALSE.}

\item{manual_ONCE_path_nlp}{Optional NLP dictionary path if O2 = FALSE.}

\item{types}{Character vector of domains (e.g., "Diagnosis", "Lab").}

\item{type_dict}{Named list mapping domains to feature prefixes.}
}
\value{
A list of plots or messages for each domain.
}
\description{
Identifies and plots ONCE-derived features most similar to the target code/CUI across domains.
}
