% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{plot_visualized_data}
\alias{plot_visualized_data}
\title{Plot Visualized Data}
\usage{
plot_visualized_data(
  data,
  count_column = NULL,
  prefix,
  description_label,
  output_path = tempdir(),
  save_plots = FALSE,
  log_scale = FALSE
)
}
\arguments{
\item{data}{A list containing two data frames: `Total_Counts` and `Patient_Counts`.}

\item{count_column}{Character. Specify whether to plot "Total_Count" or "Patient_Count". If NULL, both are plotted.}

\item{prefix}{Prefix for medical codes (e.g., "RXNORM:", "LOINC:").}

\item{description_label}{A label describing the dataset.}

\item{output_path}{Path to save the plots (default = `tempdir()`).}

\item{save_plots}{Logical. Whether to save plots as image files (default = FALSE).}

\item{log_scale}{Logical. Whether to apply a logarithmic scale to the y-axis (default = FALSE).}
}
\value{
A ggplot object or a list of ggplot objects.
}
\description{
This function generates and optionally saves histograms for patient and total counts.
}
