% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarystat.R
\name{summarystat}
\alias{summarystat}
\title{Summary statistics table}
\usage{
summarystat(data, x, y, na.rm = FALSE)
}
\value{
A table of summary statistics for each level of the given factor.
}
\description{
This function summarizes the mean, median and standard deviation of the different
levels of a factor for a given data set. NAs in the data are not removed by default
but you can change this.
}
\examples{
summarystat(palmerpenguins::penguins, "species", "body_mass_g", na.rm = TRUE)
summarystat(gapminder::gapminder, "continent", "lifeExp")
}
