% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ONCE_data.R
\name{clean_ONCE_data}
\alias{clean_ONCE_data}
\title{Load ONCE-Generated Feature Data}
\usage{
clean_ONCE_data(target_code, O2 = TRUE, path_code = NULL, path_nlp = NULL)
}
\arguments{
\item{target_code}{The codified feature ID of interest (e.g., "PheCode:335").}

\item{O2}{Logical flag for whether the script is running on O2 (default = TRUE).}

\item{path_code}{Optional file path to codified CSV (if O2 = FALSE).}

\item{path_nlp}{Optional file path to NLP CSV (if O2 = FALSE).}
}
\value{
A named list with two data frames: `code` and `nlp`, each including `feature_id`, `description`, and `target_similarity`.
}
\description{
Loads ONCE-generated codified and NLP feature dictionaries based on a target phenotype.
}
