% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{generate_intermediary_sqlite}
\alias{generate_intermediary_sqlite}
\title{Generate Intermediary SQLite Database}
\usage{
generate_intermediary_sqlite(con, output_sqlite_path, time_column = NULL)
}
\arguments{
\item{con}{A database connection object to an existing SQLite database (usually containing `df_monthly`).}

\item{output_sqlite_path}{A string indicating the file path to save the aggregated intermediary SQLite database.}

\item{time_column}{Optional. A string naming the column (e.g., "Month") from which to extract the year. The first four characters 
of each entry in this column must represent a 4-digit year (e.g., "2013", "2020-05", or "2017-12-03").}
}
\value{
Saves a new SQLite database file at the specified path containing a table named `processed_data`.
}
\description{
Aggregates patient data from a source SQLite connection and writes it to an intermediary SQLite file. 
If a `time_column` is provided (e.g., "Month"), the function will extract the year using the first four characters 
and aggregate by `Patient`, `Parent_Code`, and `Year`. Otherwise, it aggregates by `Patient` and `Parent_Code` only.
}
