% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatsFCT.R
\name{summaryStats}
\alias{summaryStats}
\title{Summary Statistics Function}
\usage{
summaryStats(dataset_of_interest, var_of_interest, export_results = FALSE)
}
\arguments{
\item{dataset_of_interest}{The dataset of interest}

\item{var_of_interest}{The variable of interest}

\item{export_results}{Logical argument, where the default argument is FALSE. If TRUE,
a csv is created within the main directory.}
}
\value{
A tibble that contains the summary statistics.
}
\description{
This summary will calculate the following summary statistics for a given variable in a
dataset of interest: mean, median, max, min, standard deviation.
}
\examples{
summaryStats(penguins, bill_length_mm)
summaryStats(mtcars, hp)

}
